fix(playwright): correct fullyQualifiedName query + relocate suite factory
The 12 SearchSeparation reindex tests and GlossaryRenameCascade's pre-rename
assertion all timed out polling /api/v1/search/query because they queried
`fullyQualifiedName.keyword:"..."`. That field is mapped as `type: keyword`
directly (with lowercase_normalizer), with no `.keyword` subfield -- the
query returned zero hits across every entity index, so the poll never saw
hits[0]._source. Sibling reindex specs already pass with the bare
`fullyQualifiedName:` form.

Also relocate searchSeparation.ts out of playwright/utils/ -- it called
test.describe.serial / test() which doesn't belong under utils -- into
playwright/e2e/Features/SearchSeparation/searchSeparationSuite.ts next to
its 12 consumer specs. Imports rewritten.
