+++ openmetadata-ui/src/main/resources/ui/playwright/e2e/Features/SearchSeparation/Pipeline.spec.ts
@@ -13,7 +13,7 @@
 
 import { test } from '@playwright/test';
 import { PipelineClass } from '../../../support/entity/PipelineClass';
-import { registerFilterSeparationSuite } from '../../../utils/searchSeparation';
+import { registerFilterSeparationSuite } from './searchSeparationSuite';
 
 test.use({ storageState: 'playwright/.auth/admin.json' });
